requirements.txt
lightgbm
shap
pandas
numpy
sqlalchemy
psycopg2-binary
fastapi
uvicorn
scikit-learn
matplotlib
seaborn
elasticsearch
jupyter
arm-pyart
boto3
